requirements.txt
certifi==2023.5.7
googleapis-common-protos==1.59.1
grpclib==0.4.5
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
multidict==6.0.4
numpy==1.25.0
opencv-contrib-python==4.5.5.62
Pillow==9.5.0
protobuf==4.23.4
requests==2.31.0
typing_extensions==4.7.1
viam-sdk==0.4.4
